test(helpers): extract shared fetch-mock setup to cut test boilerplate
Move the identical env lifecycle and response helpers into
tests/unit/helpers/fetch-mock.ts and use them from the deploy-keys and
job-token-scope registry tests. Each file keeps its own jest.mock so mock
isolation across parallel test files is unchanged.
