+++ tests/unit/entities/deploy-keys/registry.test.ts
@@ -2,58 +2,22 @@ import {
   deployKeysToolRegistry,
   getDeployKeysReadOnlyToolNames,
 } from '../../../../src/entities/deploy-keys/registry';
-import { enhancedFetch } from '../../../../src/utils/fetch';
+import {
+  installFetchMock,
+  mockOk,
+  mockNoContent,
+  lastFetchCall as lastCall,
+  mockEnhancedFetch,
+} from '../../helpers/fetch-mock';
 
 jest.mock('../../../../src/utils/fetch', () => ({
   enhancedFetch: jest.fn(),
 }));
 
-const mockEnhancedFetch = enhancedFetch as jest.MockedFunction<typeof enhancedFetch>;
-
-const originalEnv = process.env;
-
-beforeAll(() => {
-  process.env = {
-    ...originalEnv,
-    GITLAB_API_URL: 'https://gitlab.example.com',
-    GITLAB_TOKEN: 'test-token-12345',
-  };
-});
-
-afterAll(() => {
-  process.env = originalEnv;
-});
-
-beforeEach(() => {
-  jest.clearAllMocks();
-  jest.resetAllMocks();
-  mockEnhancedFetch.mockReset();
-});
-
-function mockOk(payload: unknown): void {
-  mockEnhancedFetch.mockResolvedValueOnce({
-    ok: true,
-    status: 200,
-    statusText: 'OK',
-    json: jest.fn().mockResolvedValue(payload),
-  } as unknown as Response);
-}
-
-function mockNoContent(): void {
-  mockEnhancedFetch.mockResolvedValueOnce({
-    ok: true,
-    status: 204,
-    statusText: 'No Content',
-  } as unknown as Response);
-}
+installFetchMock();
 
 const browse = () => deployKeysToolRegistry.get('browse_deploy_keys')!;
 const manage = () => deployKeysToolRegistry.get('manage_deploy_key')!;
-
-function lastCall(): [string, RequestInit | undefined] {
-  const call = mockEnhancedFetch.mock.calls.at(-1)!;
-  return [call[0], call[1]];
-}
 
 describe('Deploy Keys Registry', () => {
   describe('Registry Structure', () => {
